feat: recalibrated decision bar, Iris EIE wiring, resilience + responsiveness
- Recalibrate EIE decision bar (band on ability percentile + model-uncertainty
  floor) so Strong/Medium/Weak map to Strong Hire / Reservations / No Hire;
  verified on 10 domain-expert interviews. expertise_score column → numeric.
- Wire EIE into Iris job interviews (candidates route) + shared EieProfile
  component (admin Expert Rubrics + company job detail).
- Interviewer belief-loop client wiring (targeting, decision-stability stop,
  interview-quality) + one-score-per-turn (clientBelief flag).
- Resilience: Groq per-minute retry, DB transient-error retry + reconnect,
  clear connectivity errors at interview start.
- Voice: stop in-flight TTS before any new speech (fixes double-voice on repeat).
- Responsiveness: top NavProgress bar, disable particle canvas on app routes,
  global tactile button states; server-render the company Candidates list.

Regression suite green (scripts/test-eie.mjs).

+++ scripts/init-db.mjs
@@ -365,7 +365,7 @@ CREATE TABLE IF NOT EXISTS expert_assessments (
   domain            text,
   transcript        jsonb NOT NULL DEFAULT '[]',
   report            jsonb,
-  expertise_score   integer,
+  expertise_score   numeric,
   expertise_level   text,
   created_at        timestamptz NOT NULL DEFAULT now()
 );
